feat: add comprehensive tests for useVirtualGridList including performance, viewport calculations, and item dimensions - Implemented performance tests for handling large datasets (1,000 and 10,000 items). - Added viewport calculation tests to verify correct total height and visible range. - Created item dimension tests to ensure accurate height and width calculations based on configuration. - Removed outdated performance test file to streamline test suite.
